Create utilities.py
